Cleaning up the creative page
