+++ src/config/creative.config.js
@@ -1,5 +1,3 @@
-
-export const AnimationCardData = [
     /* 
     Each Object can take the following peramiters:  
     image: String
@@ -7,39 +5,48 @@ export const AnimationCardData = [
     description: String
     href: String
     */
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/butterfly.png', title: 'Butterfly', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://youtu.be/R1_Z7zR6drM' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/parkinglot.png', title: 'Parking Lot', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=PoLar73R4Zw' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/rainbowcar.png', title: 'Rainbow Rd', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=Fu2tjMiWqGU' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/rainbowman.png', title: 'Rainbow Man', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=TD0s8BGV2fQ&feature=youtu.be' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/strut.png', title: 'Strut Fur', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=ncLb4wvw9uY' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/wateralien.png', title: 'Fluid Alien', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=Wo40N4g04QI' },
-]
-
-export const GradshowData = [
-    /* 
-    Each Object can take the following peramiters:  
-    image: String
-    title: String
-    description: String
-    href: String
-    */
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/trumpbot.png', title: 'Trump Bot', description: 'A simulacrum of Donald Trump driven by a TensorFlow Neural Network, animated in Maya, rendered with Arnold', href: 'https://vimeo.com/317655498' },
-]
-
-export const PhotographyCardData = [
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Two+Foxes+Complete+Sides.jpg', title: 'Two Foxes', description: 'A rendering of a building concept made to learn modeling in Cinema4D', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Two+Foxes+Complete+Sides.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Three+Spikes+Final.jpg', title: 'Three Spikes', description: 'A rendering of a building concept made to learn modeling in Cinema4D', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Three+Spikes+Final.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hotel+Building.jpg', title: 'Hotel Casino', description: 'A rendering of a building concept made to learn modeling in Cinema4D', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hotel+Building.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/01+Canada+IG.jpg', title: '1 Canada Sq', description: 'An architectural rendering made using Cinema4D for the purpose of studying 3D reflections', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/01+Canada+IG.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/25+Canada+IG.jpg', title: '25 Canada Sq', description: 'An architectural rendering made using Cinema4D for the purpose of studying 3D reflections', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/25+Canada+IG.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/BT+Tower.jpg', title: 'BT Tower', description: 'An architectural rendering made using Cinema4D for the purpose of studying 3D reflections', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/BT+Tower.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hearing_Ear.jpg', title: 'Sine Ear', description: 'A rendering of the human anatomy represented by data made with ProcessingPY, lazer cutter, darkroom exposures', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hearing_Ear.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Sight_Eye.jpg', title: 'Sine Eye', description: 'A rendering of the human anatomy represented by data made with ProcessingPY, lazer cutter, darkroom exposures', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Sight_Eye.jpg' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Touch_finger.jpg', title: 'Sine Finger', description: 'A rendering of the human anatomy represented by data made with ProcessingPY, lazer cutter, darkroom exposures', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Touch_finger.jpg' },
-]
-
-export const WebDesignCardData = [
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/LoFi+fintech+style+dashboard+Screenshot.png', title: 'FinTech Style Dash Wireframes', description: 'Low Fideliy Wire Frames to guide the design and implimentation of a FinTech style personal expenses dashboard Proof of Concept, made in Sketch', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/LoFi+FinTech+Dashboard+01+wireframes.png' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/LoFi+jamescrookdev+Screenshot.png', title: 'jamescrookdev Wireframes', description: 'Low Fideliy Wire Frames to guide the design of this website, made in Sketch', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/my-react-portfolio+design+01.pdf' },
-    { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/HIFI+jamescrookdev+Screenshot.png', title: 'jamescrookdev HIFI Design', description: 'High Fidelity Designs to guide the construction of this website, made in Sketch', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/my-react-portfolio+HiFi+01.pdf' },
+export const Data = [
+    {
+        title: "Web Design",
+        id: "WebDesignCardData",
+        content: [
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/LoFi+fintech+style+dashboard+Screenshot.png', title: 'FinTech Style Dash Wireframes', description: 'Low Fideliy Wire Frames to guide the design and implimentation of a FinTech style personal expenses dashboard Proof of Concept, made in Sketch', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/LoFi+FinTech+Dashboard+01+wireframes.png' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/LoFi+jamescrookdev+Screenshot.png', title: 'jamescrookdev Wireframes', description: 'Low Fideliy Wire Frames to guide the design of this website, made in Sketch', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/my-react-portfolio+design+01.pdf' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/HIFI+jamescrookdev+Screenshot.png', title: 'jamescrookdev HIFI Design', description: 'High Fidelity Designs to guide the construction of this website, made in Sketch', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/my-react-portfolio+HiFi+01.pdf' },
+        ]
+    },
+    {
+        title: "Animations",
+        id: "AnimationCardData",
+        content: [
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/butterfly.png', title: 'Butterfly', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://youtu.be/R1_Z7zR6drM' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/parkinglot.png', title: 'Parking Lot', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=PoLar73R4Zw' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/rainbowcar.png', title: 'Rainbow Rd', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=Fu2tjMiWqGU' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/rainbowman.png', title: 'Rainbow Man', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=TD0s8BGV2fQ&feature=youtu.be' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/strut.png', title: 'Strut Fur', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=ncLb4wvw9uY' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/wateralien.png', title: 'Fluid Alien', description: 'An animation made using Cinema4D, Adobe Mixamo, AdobeFuse, Adobe AfterEffects', href: 'https://www.youtube.com/watch?v=Wo40N4g04QI' },
+        ]
+    },
+    {
+        title: "Grad Show",
+        id: "GradshowData",
+        content: [
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/trumpbot.png', title: 'Trump Bot', description: 'A simulacrum of Donald Trump driven by a TensorFlow Neural Network, animated in Maya, rendered with Arnold', href: 'https://vimeo.com/317655498' },
+        ]
+    },
+    {
+        title: "Photography",
+        id: "PhotographyCardData",
+        content: [
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Two+Foxes+Complete+Sides.jpg', title: 'Two Foxes', description: 'A rendering of a building concept made to learn modeling in Cinema4D', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Two+Foxes+Complete+Sides.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Three+Spikes+Final.jpg', title: 'Three Spikes', description: 'A rendering of a building concept made to learn modeling in Cinema4D', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Three+Spikes+Final.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hotel+Building.jpg', title: 'Hotel Casino', description: 'A rendering of a building concept made to learn modeling in Cinema4D', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hotel+Building.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/01+Canada+IG.jpg', title: '1 Canada Sq', description: 'An architectural rendering made using Cinema4D for the purpose of studying 3D reflections', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/01+Canada+IG.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/25+Canada+IG.jpg', title: '25 Canada Sq', description: 'An architectural rendering made using Cinema4D for the purpose of studying 3D reflections', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/25+Canada+IG.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/BT+Tower.jpg', title: 'BT Tower', description: 'An architectural rendering made using Cinema4D for the purpose of studying 3D reflections', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/BT+Tower.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hearing_Ear.jpg', title: 'Sine Ear', description: 'A rendering of the human anatomy represented by data made with ProcessingPY, lazer cutter, darkroom exposures', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Hearing_Ear.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Sight_Eye.jpg', title: 'Sine Eye', description: 'A rendering of the human anatomy represented by data made with ProcessingPY, lazer cutter, darkroom exposures', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Sight_Eye.jpg' },
+            { image: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Touch_finger.jpg', title: 'Sine Finger', description: 'A rendering of the human anatomy represented by data made with ProcessingPY, lazer cutter, darkroom exposures', href: 'https://jamescrookdev.s3.eu-west-2.amazonaws.com/images/Touch_finger.jpg' },
+        ]
+    },
 ]
